AI-INTERVIEWER/requirement.txt
livekit
livekit-agents
livekit-plugins-deepgram
livekit-plugins-google
livekit-plugins-silero
python-dotenv
deepgram 
google-generativeai